src/agriha/vps/requirements.txt
line-bot-sdk>=3.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
nest_asyncio>=1.5
anthropic>=0.40.0
astral>=3.2
pyyaml>=6.0
